requirements.txt
asgiref==3.2.10
backports.zoneinfo==0.2.1
dj-database-url==1.0.0
Django==3.1
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-filter==2.3.0
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.1.0
Markdown==3.2
Pillow==9.3.0
psycopg2==2.9.5
pytz==2022.5
sqlparse==0.4.3
tzdata==2022.6
whitenoise==6.2.0
